requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2026.1.4
    # via
    #   kubernetes
    #   requests
charset-normalizer==3.4.4
    # via requests
chartpress==2.3.0
    # via -r requirements.in
docker==7.1.0
    # via chartpress
durationpy==0.10
    # via kubernetes
exceptiongroup==1.3.1
    # via pytest
execnet==2.1.2
    # via pytest-xdist
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
kubernetes==35.0.0
    # via -r requirements.in
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   pytest
    #   pytest-rerunfailures
pluggy==1.6.0
    # via pytest
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-rerunfailures
    #   pytest-timeout
    #   pytest-xdist
pytest-rerunfailures==16.1
    # via -r requirements.in
pytest-timeout==2.4.0
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via kubernetes
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   kubernetes
requests==2.32.5
    # via
    #   -r requirements.in
    #   docker
    #   kubernetes
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
ruamel-yaml==0.19.1
    # via chartpress
six==1.17.0
    # via
    #   kubernetes
    #   python-dateutil
tomli==2.4.0
    # via pytest
typing-extensions==4.15.0
    # via exceptiongroup
urllib3==2.6.3
    # via
    #   docker
    #   kubernetes
    #   requests
websocket-client==1.9.0
    # via kubernetes
